Moved graphRegistry to graphManager and created seperate graphRegistry.
